requirements.txt
# ML tools
numpy
pandas
joblib

# System tools
fabric
invoke
twine

# hyperparameters
parameter-sherpa

pyyaml